Exception handling - delete

+++ src/main/java/com/example/course/services/UserService.java
@@ -1,15 +1,19 @@
 package com.example.course.services;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.course.entities.User;
 import com.example.course.repositories.UserRepository;
+import com.example.course.services.exceptions.DatabaseException;
 import com.example.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -35,7 +39,17 @@ public class UserService {
 	
 	@Transactional
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if(userRepository.existsById(id)) {
+				userRepository.deleteById(id);
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+		
 	}
 	
 	@Transactional
